refactor shop folders
