+++ hamza-client/src/modules/shop/components/desktop-side-filter/store-side-menu.tsx
@@ -5,7 +5,7 @@ import { Box, Text, Heading, Flex, Skeleton } from '@chakra-ui/react';
 import CategoryButton from './category-button';
 import ReviewButton from '../review-button';
 import FilterButton from './filter-button';
-import RangeSlider from '../range-slider';
+import RangeSlider from './range-slider';
 import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
