+++ hamza-client/src/modules/shop/components/filter-carousel/filter-bar-store.tsx
